thread criticality (#36)
* add SchedulingPolicy enum and ProgramGroup policy builder option

This commit adds a SchedulingPolicy for ProgramGroups, which gets passed down
to ProgramVersions as they are loaded. The SchedulingPolicy can be any valid Linux
scheduling policy, as defined and documented in the enum. The policy gets passed down
to the `perf_map_poller` thread, which sets its policy and priority (and niceness, if
relevant) appropriately.

Note: this currently uses a patched version of the `thread-priority` crate, until 
SCHED_DEADLINE support is merged in.